Split PrimaryTaxonomyPicker into container and component
